Push notification fixes

+++ Assets/Scripts/VideoProduction/EnergyManager.cs
@@ -36,6 +36,7 @@ public class EnergyManager : MonoBehaviour
     bool isResting;
     bool energyChargeBlocked = false;
     [Inject] private IPushNotificationsManager pushNotifications;
+    [SerializeField] private int energyNotificationThreshoold = 5;
 
     // Start is called before the first frame update
     void Start()
@@ -206,24 +207,27 @@ public class EnergyManager : MonoBehaviour
 
     void OnApplicationFocus(bool hasFocus)
     {
-        if(!hasFocus)
+        if(GetEnergy () < GetMaxEnergy() - energyNotificationThreshoold)
         {
-            SaveEnergyData ();
-            var title = "Energy Full";
-            var subtitle = "Login to create a video.";
-            var text = new string[]{"Ready to create another masterpiece?", "You're ready to create another video.","You did a good job resting, it's time to get back to work again!"};
-            var timeTrigger = SecondsToRefillEnergy();
-            var id = 1;
-            pushNotifications.ScheduleNotification(title, subtitle, text[UnityEngine.Random.Range(0, text.Length)], timeTrigger, id);
-            PlayerPrefs.SetInt("EnergyNotification", id);
-        }
-        else
-        {
-            if(PlayerPrefs.HasKey("EnergyNotification"))
+            if(!hasFocus)
             {
-                var id = PlayerPrefs.GetInt("EnergyNotification");
-                pushNotifications.UnScheduleNotification(id);
-                PlayerPrefs.DeleteKey("EnergyNotification");
+                SaveEnergyData ();
+                var title = "Energy Full";
+                var subtitle = "Login to create a video.";
+                var text = new string[]{"Ready to create another masterpiece?", "You're ready to create another video.","You did a good job resting, it's time to get back to work again!"};
+                var timeTrigger = SecondsToRefillEnergy();
+                var id = 1;
+                pushNotifications.ScheduleNotification(title, subtitle, text[UnityEngine.Random.Range(0, text.Length)], timeTrigger, id);
+                PlayerPrefs.SetInt("EnergyNotification", id);
+            }
+            else
+            {
+                if(PlayerPrefs.HasKey("EnergyNotification"))
+                {
+                    var id = PlayerPrefs.GetInt("EnergyNotification");
+                    pushNotifications.UnScheduleNotification(id);
+                    PlayerPrefs.DeleteKey("EnergyNotification");
+                }
             }
         }
     }
